face-service/requirements.txt
flask
pymongo
python-dotenv
face_recognition
numpy
Pillow
gunicorn